Create additionalCheckoutOptions.js
This function is called and looks from the OrderFinalStep page of uStore. Use this if you want to add a custom field to your checkout that will show in Special Instructions. It uses a custom dropdown that is prepended to the special instructions. Even if the user comes back to the product, it separates out the special instructions and the custom field. This using jQuery pretty heavily, so your theme will need to call it (if it doesn't already).